requirements.txt
uvicorn
fastapi
opencv-python
numpy
requests
Pillow
torch
torchvision
transformers